+++ src/app.ts
@@ -1,5 +1,5 @@
 import express, { Application} from 'express';
-import apiRouter from './controllers/api';
+import apiRouter from './routes/routes';
 
 const app: Application = express();
 app.use('/v1/api', apiRouter)
